Added a document on how to create source code directories for games.
The conversation was originally from the Ludum Dare offshoot community called #wizardslair on the irc network Afternet.org.

Document records how to recreate Blackbird's directory structure from scratch when a new project is required.

Committed this advice because a record of how to start a new project would be helpful in the future.